+++ src/Oxrun/Application/oxrun-light.php
@@ -9,7 +9,34 @@
 
 declare(strict_types=1);
 
-$application = new \Symfony\Component\Console\Application('oxrun-light', '0.2');
+namespace Oxrun\Application;
+
+/**
+ * Class OxrunLight
+ * @package Oxrun\Application
+ */
+class OxrunLight extends \Symfony\Component\Console\Application
+{
+    /**
+     * @inheritDoc
+     */
+    public function getHelp()
+    {
+        $version = parent::getHelp();
+        return <<<TAG
+$version
+
+-----------------------------------------------------------------------
+<comment>  Is a light console line tool for OXID eSale.
+  These commands don't need an active OXID eSale database connection.</comment>
+-----------------------------------------------------------------------
+TAG;
+    }
+
+}
+
+$application = new OxrunLight('oxrun-light', '0.3');
+$application->add(new \Oxrun\Command\Cache\ClearCommand());
 $application->add(new \Oxrun\Command\Misc\RegisterCommand());
 $application->add(new \Oxrun\Command\Misc\PhpstormMetadataCommand());
 $application->add(new \Oxrun\Command\Misc\GenerateDocumentationCommand());