+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [v5.1.0] 2021-09-07
+### Changed
+  - `oxrun-light` is now officially documented.
+  - Command cache:clear works without bootstrap.php so can now used in `oxrun-light`.
 
 ## [v5.0.1] 2021-09-06
 ### Fixed